pyproject.toml
[project]
name = "tune_xvf_3800"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scipy",
  "soundfile",
  "pyyaml",
  "matplotlib",
]

[project.scripts]
tune-xvf = "tune_xvf_3800.cli:main"
